+++ common/src/main/java/cc/cassian/cauldrons/compat/jade/CauldronComponentProvider.java
@@ -3,15 +3,17 @@ package cc.cassian.cauldrons.compat.jade;
 import cc.cassian.cauldrons.blocks.BrewingCauldronBlock;
 import cc.cassian.cauldrons.blocks.entity.CauldronBlockEntity;
 import net.minecraft.ChatFormatting;
-import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.item.Items;
-import net.minecraft.world.item.alchemy.PotionContents;
+import net.minecraft.world.level.block.CauldronBlock;
 import snownee.jade.api.BlockAccessor;
 import snownee.jade.api.IBlockComponentProvider;
 import snownee.jade.api.ITooltip;
+import snownee.jade.api.JadeIds;
 import snownee.jade.api.config.IPluginConfig;
+import snownee.jade.api.theme.IThemeHelper;
+
+import java.util.Locale;
 
 public enum CauldronComponentProvider implements IBlockComponentProvider {
     INSTANCE;
@@ -19,21 +21,10 @@ public enum CauldronComponentProvider implements IBlockComponentProvider {
     @Override
     public void appendTooltip(ITooltip iTooltip, BlockAccessor blockAccessor, IPluginConfig iPluginConfig) {
         if (blockAccessor.getBlockEntity() instanceof CauldronBlockEntity cauldronBlockEntity) {
-            if (!cauldronBlockEntity.getItem().isEmpty()) {
-                iTooltip.add(cauldronBlockEntity.getItem().getHoverName());
-                iTooltip.add(Component.empty());
+            for (Component component : cauldronBlockEntity.getForWaila(blockAccessor.getBlockState())) {
+                iTooltip.add(component);
             }
-            if (cauldronBlockEntity.getPotion() != PotionContents.EMPTY) {
-                iTooltip.add(Component.translatable("gui.toil_and_trouble.doses", blockAccessor.getBlockState().getValue(BrewingCauldronBlock.POTION_QUANTITY)).withStyle(ChatFormatting.DARK_PURPLE));
-                var item = Items.POTION;
-                if (cauldronBlockEntity.isPotionSplash())
-                    item = Items.SPLASH_POTION;
-                else if (cauldronBlockEntity.isPotionLingering())
-                    item = Items.LINGERING_POTION;
-                iTooltip.add(CauldronBlockEntity.createItemStack(item, cauldronBlockEntity.getPotion()).getHoverName());
-                if (Screen.hasShiftDown())
-                    PotionContents.addPotionTooltip(cauldronBlockEntity.getPotion().getAllEffects(), iTooltip::add, 0, 0);
-            }
+            iTooltip.replace(JadeIds.CORE_OBJECT_NAME, Component.translatable("block.toil_and_trouble.cauldron." + blockAccessor.getBlockState().getValue(BrewingCauldronBlock.CONTENTS).name().toLowerCase(Locale.ROOT)).withStyle(ChatFormatting.WHITE));
         }
     }