+++ common/src/main/java/cc/cassian/cauldrons/recipe/BrewingRecipeInput.java
@@ -1,17 +1,19 @@
 package cc.cassian.cauldrons.recipe;
 
+import cc.cassian.cauldrons.core.CauldronContents;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.alchemy.PotionContents;
 import net.minecraft.world.item.crafting.RecipeInput;
 
 public class BrewingRecipeInput implements RecipeInput {
 
     private final ItemStack reagent;
-    private final PotionContents potion;
+    private final CauldronContents contents;
+    private final boolean isHeated;
 
-    public BrewingRecipeInput(ItemStack reagent, PotionContents potion) {
+    public BrewingRecipeInput(ItemStack reagent, CauldronContents contents, boolean isHeated) {
         this.reagent = reagent;
-        this.potion = potion;
+        this.contents = contents;
+        this.isHeated = isHeated;
     }
 
     @Override
@@ -19,12 +21,16 @@ public class BrewingRecipeInput implements RecipeInput {
         return reagent;
     }
 
-    public PotionContents getPotionContents() {
-        return potion;
+    public CauldronContents getContents() {
+        return contents;
     }
 
     @Override
     public int size() {
         return 2;
+    }
+
+    public boolean isHeated() {
+        return isHeated;
     }
 }